requirements.txt
ag-ui-protocol==0.1.8
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.61.0
anyio==4.10.0
argcomplete==3.6.2
astor==0.8.1
async-timeout==5.0.1
attrdict==2.0.1
attrs==25.3.0
babel==2.17.0
bce-python-sdk==0.9.42
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.40.2
botocore==1.40.2
cachetools==5.3.1
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
cohere==5.16.2
colorama==0.4.6
contourpy==1.3.2
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.1.2
decorator==5.2.1
distro==1.9.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.103.1
fastavro==1.12.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-babel==4.0.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
google-auth==2.40.3
google-genai==1.28.0
griffe==1.9.0
groq==0.30.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.3
idna==3.10
imageio==2.37.0
imgaug==0.4.0
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.2
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
lmdb==1.7.3
logfire-api==4.1.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mcp==1.12.3
mdurl==0.1.2
mistralai==1.9.3
more-itertools==10.7.0
multidict==6.6.3
networkx==3.4.2
numpy==1.24.4
openai==1.99.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.36.0
opt-einsum==3.3.0
packaging==25.0
paddle-bfloat==0.1.7
paddleocr==2.6.1.3
paddlepaddle==2.5.0
pandas==2.3.1
pdf2docx==0.5.8
pdf2image==1.16.3
Pillow==10.0.1
pluggy==1.6.0
premailer==3.10.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-ai==0.5.0
pydantic-ai-slim==0.5.0
pydantic-evals==0.5.0
pydantic-graph==0.5.0
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.20.2
pyparsing==3.2.3
pytest==7.4.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
rarfile==4.2
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.1
scikit-image==0.24.0
scipy==1.15.3
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sse-starlette==3.0.2
starlette==0.47.2
tenacity==8.5.0
termcolor==3.1.0
tifffile==2025.5.10
tokenizers==0.21.4
tomli==2.2.1
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.23.2
visualdl==2.5.3
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
